requirements.txt
requests
pandas
scikit-learn
python-dotenv
matplotlib
streamlit
prophet
dedalus_labs
